functions/report_execution/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
cycler==0.10.0
google-api-core==1.31.0
google-auth==1.33.0
google-cloud==0.34.0
google-cloud-core==1.7.1
google-cloud-storage==1.41.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
greenlet==1.1.0
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.2
mysql-connector-python==8.0.25
numpy==1.21.0
packaging==21.0
pandas==1.3.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.21
urllib3==1.26.6
